feat: add warehouse & granary capacity depends on their level

+++ src/main/java/io/lanu/travian/templates/buildings/GranaryBuilding.java
@@ -5,13 +5,15 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = true)
 @Data
 @NoArgsConstructor
 public class GranaryBuilding extends BuildingBase {
-    private Integer capacity;
+    private BigDecimal capacity;
 
-    public GranaryBuilding(Integer capacity) {
+    public GranaryBuilding(BigDecimal capacity) {
         super();
         this.name = EBuildings.GRANARY.getName();
         this.capacity = capacity;
