+++ src/main/java/io/lanu/travian/templates/buildings/WarehouseBuilding.java
@@ -5,13 +5,15 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @EqualsAndHashCode(callSuper = true)
 @Data
 @NoArgsConstructor
 public class WarehouseBuilding extends BuildingBase{
-    private Integer capacity;
+    private BigDecimal capacity;
 
-    public WarehouseBuilding(Integer capacity) {
+    public WarehouseBuilding(BigDecimal capacity) {
         super();
         this.name = EBuildings.WAREHOUSE.getName();
         this.capacity = capacity;
